Rename handler.mjs to http.mjs
